feat: merge `web` and `backend_web` modules
